refactor itemstable hooks, update invitation type, clean up imports
