+++ context/app/static/js/components/workspaces/ConfirmDeclineInvitationDialog/ConfirmDeclineInvitationDialog.tsx
@@ -8,6 +8,7 @@ import { useWorkspaceToasts } from 'js/components/workspaces/toastHooks';
 import { useWorkspacesEventContext } from 'js/components/workspaces/contexts';
 import { trackEvent } from 'js/helpers/trackers';
 import { WorkspacesEventCategories } from 'js/components/workspaces/types';
+import { getSharerInfo } from 'js/components/workspaces/utils';
 
 export default function ConfirmDeclineInvitationDialog() {
   const { handleDeleteInvitation } = useInvitationsList();
@@ -22,11 +23,8 @@ export default function ConfirmDeclineInvitationDialog() {
   const {
     shared_workspace_id: { id, name },
   } = invitation;
-  const {
-    original_workspace_id: {
-      user_id: { first_name, last_name },
-    },
-  } = invitation;
+
+  const { first_name, last_name } = getSharerInfo(invitation);
 
   const trackDialogEvent = (actionSuffix: string) => {
     const action =
